Erro menor

+++ app/Observers/TaskUserObserver.php
@@ -47,7 +47,9 @@ class TaskUserObserver
         $log->users()->save($user);
 
 		// Remover os votos do usuário ao sair da task
-		$user->votes()->detach($task);
+		if ($user->votes->contains($task)) {
+			$user->votes()->detach($task);
+		}
 
 		// Recontar os votos, já que a equipe diminuiu
         if ($task->votes()->count() >= $task->users()->count() / 2) {
